refactor(math): split ComputeStats cluster into TrajectoryMath.Stats.cs
